+++ integrations/checkmarx-one/CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
-- Added webhook support for the KICS Scan Result
+- Added webhook support for the KICS and SAST Scan Result
 
 
 ## 0.2.2-dev (2025-09-08)